Add new search option to ICOS_sr_search
Added a search with a fixed set of choices for ICOS mirrors
while allowing the length to vary. This may be useful for
investigating possible configurations using existing optics.

Also included changes to allow explicit specification of
the index of refraction, which is necessary when considering
alternate materials or wavelengths.